+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 ### Added
 - `contour` function
+- `Polygon.transform` to apply a general transformation to the polygon vertices
 - `Polygon.contain` tests whether single points are inside the polygon
 - `Polygon.contain_all` and `Polygon.contain_any` test multiple points with short circuit
 - `all_inside` and `any_inside` test multiple points against multiple polygons with short circuit